+++ packages/plinth_charts/lib/plinth_charts.dart
@@ -12,4 +12,5 @@ library;
 
 export 'package:plinth_core/plinth_core.dart';
 
+export 'src/plinth_line_chart.dart';
 export 'src/plinth_sparkline.dart';